Fix build for x64 arch
media_info_get_size waits unsigned long long not uint64_t
even if it seems to be the same in 32 bits arch.

=> int media_info_get_size(media_info_h media, unsigned long long *size)

log:

[   27s] content/content_instance.cc: In member function 'void ContentItem::init(media_info_h)':
[   27s] content/content_instance.cc:313:38: error: invalid conversion from 'uint64_t* {aka long unsigned int*}' to 'long long unsigned int*' [-fpermissive]
[   27s]    if (media_info_get_size(handle, &ll) == MEDIA_CONTENT_ERROR_NONE)
[   27s]                                       ^

BUG=https://crosswalk-project.org/jira/browse/XWALK-1346

+++ content/content_instance.cc
@@ -309,7 +309,7 @@ void ContentItem::init(media_info_h handle) {
   if (media_info_get_rating(handle, &i) == MEDIA_CONTENT_ERROR_NONE)
     setRating(i);
 
-  uint64_t ll;
+  unsigned long long ll; // NOLINT
   if (media_info_get_size(handle, &ll) == MEDIA_CONTENT_ERROR_NONE)
     setSize(ll);
